Add files via upload
A small company Axon, which is retailer selling cars, is facing issues in managing and analyzing their sales data. The sales data. The sales team is struggling to make sense of the data and they do not have centralize system to manage and analyze the data. The management is unable to get accurate and up-to-date, sales reports, which is affecting the decision making process. By using power bi and MySql the centralized dashboard is created and the KPI's are  identified which make the easy to access the data and get insights.